requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pillow==10.3.0
python-multipart==0.0.9
sse-starlette==2.1.0
